+++ source/extensions/omni.isaac.orbit_tasks/omni/isaac/orbit_tasks/locomotion/velocity/velocity_env_cfg.py
@@ -30,8 +30,7 @@ import omni.isaac.orbit_tasks.locomotion.velocity.mdp as mdp
 ##
 # Pre-defined configs
 ##
-# isort: off
-from omni.isaac.orbit.terrains.config.rough import ROUGH_TERRAINS_CFG
+from omni.isaac.orbit.terrains.config.rough import ROUGH_TERRAINS_CFG  # isort: skip
 
 
 ##